+++ KablunkEngine/src/Platform/Vulkan/VulkanMaterial.cpp
@@ -219,7 +219,7 @@ namespace Kablunk
 				}
 
 				pending_descriptor->write_descriptor_set.pImageInfo = array_image_infos.data();
-				pending_descriptor->write_descriptor_set.descriptorCount = array_image_infos.size();
+				pending_descriptor->write_descriptor_set.descriptorCount = static_cast<uint32_t>(array_image_infos.size());
 				m_write_descriptors[frame_index].push_back(pending_descriptor->write_descriptor_set);
 			}
 		}
